backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
alembic==1.14.1
SQLAlchemy==2.0.38
pyodbc==5.2.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.20
